Implemented state design pattern to reduce the size of the if else statements
